[googlemaps] Extract event.d.ts
