+++ types/googlemaps/tsconfig.json
@@ -19,6 +19,7 @@
     },
     "files": [
         "index.d.ts",
+        "reference/event.d.ts",
         "googlemaps-tests.ts"
     ]
 }
